feat: Use directly the ota file without extracting payload file
A device wastes memory when it applies an A/B update. First it downloads
the ota; then extracts and stores to 'data' the payload file and finally
it delegates the update_engine to apply the payload.
Now, to reduce
the needed size, the ota file is passed to the update engine without
payload file is extracted.

Signed-off-by: Daniele Sergio <daniele.sergio@kynetics.com>