fix(Service): client sends the last_log to the server without rebooting
Single copy: in case of signature errors during an OTA update, the client
sends the last_log to the server.
However, if the signature is wrong the update is stopped before rebooting
into recovery, so the last_log is not actually updated and it is not
related to this update.
The last_log should be sent to the server only when the device reboots
in recovery mode.

Closes UF-192

Signed-off-by: Daniele Sergio <daniele.sergio@kynetics.com>