fix: service reboot on configuration change
The process that verifies if the service must be rebooted when
the configuration is changed is wrongly implemented.
The  UFService overwrites the current configuration with the
new one before to search the differences with the current
configuration so basically it checks differences between two new
configurations that are always equal.
Fixed by searching differences between the new configuration
and the old one.

Signed-off-by: Daniele Sergio <daniele.sergio@kynetics.com>