feat: store target token received by the update server  in a new property
TargetToken received by the update server is stored in a new property.
When the uf configuration is loaded by the shared preferences, the
target token to use is selected with the following priority :
 1- (high priority) target token provided by a manual configuration
    (ui/message/configuration file)
 2- targetToken received by update server
 3- (lower priority) an empty string

The target token received from the update server is cleared when a
configuration with a different controller id or tenant is provided.